perf(test): split ssh.device-config.test.ts along its describe boundaries
